Rename focus to navigation to be more generic
