+++ src/views/Home.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect, useRef } from 'react'
 import { useHistory } from 'react-router-dom'
 import { getTrendingMovies, getTrendingShows } from '../services/trending'
 import { MediaModel } from '../models/interfaces/MediaModel'
-import { Section } from '../navigation/Section'
+import { Section } from '../components/Section'
 import { ListRow } from '../components/ListRow'
 import { Featured } from '../components/Featured'
 import { ListWrapper } from '../components/ListWrapper'
